Refactor: SRP에 맞게 컴포넌트 분리
